requirements.txt
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-identity
paramiko
python-dotenv
flask
flask-cors
flask-smorest
python-dotenv